% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arousal.R
\name{get_arousal}
\alias{get_arousal}
\title{Get Arousal}
\usage{
get_arousal(text, long = FALSE)
}
\arguments{
\item{text}{A string}

\item{long}{A logical indicates which lexicon to use}
}
\value{
The valence of \code{text}
}
\description{
Calculates the mean arousal of hebrew text. The function can use a small lexicon (22k tokens, default) or a large lexicon (1.4M tokens)
}
\examples{
get_arousal("some Hebrew Text", long=FALSE)
get_arousal("some Hebrew Text", long=TRUE)

}
\references{
{
  \insertRef{Armony-Sivan2013-zx}{hebnoRms}

  \insertRef{joulin2016fasttext}{hebnoRms}

  \insertRef{Silber-Varod2017-xp}{hebnoRms}
}
}
\author{
Almog Simchon, \email{almogsi@post.bgu.ac.il}
}
